+++ cmd/run/show.go
@@ -6,24 +6,12 @@ import (
 	"errors"
 	"log/slog"
 	"strconv"
-	"time"
 
 	"github.com/samcarswell/trochilus/config"
 	"github.com/samcarswell/trochilus/core"
 	"github.com/samcarswell/trochilus/opts"
 	"github.com/spf13/cobra"
 )
-
-type RunShow struct {
-	ID            int64
-	CronName      string
-	StartTime     time.Time
-	EndTime       time.Time
-	LogFile       string
-	SystemLogFile string
-	Status        string
-	Duration      string
-}
 
 var showCmd = &cobra.Command{
 	Use:   "show",
@@ -32,6 +20,7 @@ var showCmd = &cobra.Command{
 		logger := slog.Default()
 		runId := opts.GetInt64OrExit(cmd, "run-id")
 		queries := config.GetDatabase(cmd.Context())
+		conf := config.GetConfig()
 
 		runRow, err := queries.GetRun(context.Background(), runId)
 		if err != nil {
@@ -41,17 +30,17 @@ var showCmd = &cobra.Command{
 				core.LogErrorAndExit(logger, err)
 			}
 		}
-		data := RunShow{
+		data := core.RunShow{
 			ID:            runRow.Run.ID,
 			CronName:      runRow.Cron.Name,
-			StartTime:     runRow.Run.StartTime,
-			EndTime:       runRow.Run.EndTime.Time,
+			StartTime:     core.FormatTime(runRow.Run.StartTime, conf.LocalTime),
+			EndTime:       core.FormatTime(runRow.Run.EndTime.Time, conf.LocalTime),
 			LogFile:       runRow.Run.LogFile,
 			SystemLogFile: runRow.Run.ExecLogFile,
 			Status:        runRow.Run.Status,
 		}
 		if runRow.Run.EndTime.Valid {
-			data.Duration = data.EndTime.Sub(data.StartTime).String()
+			data.Duration = runRow.Run.EndTime.Time.Sub(runRow.Run.StartTime).String()
 		}
 		core.PrintJson(data)
 	},